Changed: ingredient database type of macros to decimal

+++ Data/Ingredient.cs
@@ -6,8 +6,8 @@ namespace TrainingPlanApp.Web.Data
 	{
 		public int Id { get; set; }
 		public string? Name { get; set; }
-		public int? Proteins { get; set; }
-		public int? Carbohydrates { get; set; }
-		public int? Fats { get; set; }
+		public decimal? Proteins { get; set; }
+		public decimal? Carbohydrates { get; set; }
+		public decimal? Fats { get; set; }
 	}
 }
